src/requirements.txt
beautifulsoup4==4.11.1
boto3==1.26.41
botocore==1.29.41
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-lib==0.1.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
slack-sdk==3.19.5
soupsieve==2.3.2.post1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
